fix: separated homepage logic from footer logic, imported fonts
